Run filtering is now shared: runs page and job runs page
